backend/auth/requirements.txt
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.2
pydantic==2.5.0
